backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pandas==2.2.0
pandas-ta
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
